+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 1e6b744 - v13
+| Author: fjrodafo - Wed, 11 Dec 2024 16:51:31 +0000
+| 
 * Commit: 0cc96aa - v12
 | Author: fjrodafo - Sat, 7 Dec 2024 19:05:06 +0000
 |